Revert "Fix swith-case"
This reverts commit 52408d366c6d03f51b0fba373504a6793906b701.

+++ fuzzer/src/shm_conds.rs
@@ -29,7 +29,7 @@ cond: shm::SHM::<CondStmtBase>::new(),
     self.cond.cmpid = cmpid;
     self.cond.context = ctx;
     self.cond.order= order;
-    self.cond.condition = std::u64::MAX;
+    self.cond.condition = std::u32::MAX;
   }
 
 }
